fix(vllm): keep WorkerActor importable without vllm for multi-GPU launch
